requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
openpyxl==3.1.5
matplotlib==3.9.2
Werkzeug==3.0.4
numpy==2.1.1
gunicorn==23.0.0
supabase==2.10.0
python-dotenv==1.0.0